+++ crates/swc_ecma_minifier/src/analyzer/base.rs
